+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Changes to the GraphQL package are listed here. Releases follow semantic
 versioning.
 
+## [1.2.7] - 2021-01-16
+
+### Fixed
+- Validation bug when checking an object's list/non-null field argument matches
+  the type defined by the interface the object is implementing.
+
 ## [1.2.6] - 2021-01-07
 
 ### Fixed